requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.18
certifi==2019.9.11
chardet==3.0.4
Django==2.2.13
django-cors-headers==3.1.0
djangorestframework==3.10.3
idna==2.8
kombu==3.0.37
psycopg2-binary==2.8.2
pytz==2019.2
redis==2.10.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
